Meteor.settings.cfs.worker.enabled 控制worker observe

+++ packages/cfs-worker/fileWorker.js
@@ -19,38 +19,42 @@ var fs = Npm.require('fs');
  * Sets up observes on the fsCollection to store file copies and delete
  * temp files at the appropriate times.
  */
-FS.FileWorker.observe = function(fsCollection) {
+FS.FileWorker.observe = function (fsCollection) {
 
-  // Initiate observe for finding newly uploaded/added files that need to be stored
-  // per store.
-  FS.Utility.each(fsCollection.options.stores, function(store) {
-    var storeName = store.name;
-    fsCollection.files.find(getReadyQuery(storeName), {
-      fields: {
-        copies: 0
-      }
-    }).observe({
-      added: function(fsFile) {
-        // added will catch fresh files
-        FS.debug && console.log("FileWorker ADDED - calling saveCopy", storeName, "for", fsFile._id);
-        saveCopy(fsFile, storeName);
-      },
-      changed: function(fsFile) {
-        // changed will catch failures and retry them
-        FS.debug && console.log("FileWorker CHANGED - calling saveCopy", storeName, "for", fsFile._id);
-        saveCopy(fsFile, storeName);
+  if (Meteor.settings.cfs && Meteor.settings.cfs.worker && Meteor.settings.cfs.worker.enabled) {
+    // Initiate observe for finding newly uploaded/added files that need to be stored
+    // per store.
+    FS.Utility.each(fsCollection.options.stores, function (store) {
+      var storeName = store.name;
+      fsCollection.files.find(getReadyQuery(storeName), {
+        fields: {
+          copies: 0
+        }
+      }).observe({
+        added: function (fsFile) {
+          // added will catch fresh files
+          FS.debug && console.log("FileWorker ADDED - calling saveCopy", storeName, "for", fsFile._id);
+          saveCopy(fsFile, storeName);
+        },
+        changed: function (fsFile) {
+          // changed will catch failures and retry them
+          FS.debug && console.log("FileWorker CHANGED - calling saveCopy", storeName, "for", fsFile._id);
+          saveCopy(fsFile, storeName);
+        }
+      });
+    });
+
+    // Initiate observe for finding files that have been stored so we can delete
+    // any temp files
+    fsCollection.files.find(getDoneQuery(fsCollection.options.stores)).observe({
+      added: function (fsFile) {
+        FS.debug && console.log("FileWorker ADDED - calling deleteChunks for", fsFile._id);
+        FS.TempStore.removeFile(fsFile);
       }
     });
-  });
+  }
 
-  // Initiate observe for finding files that have been stored so we can delete
-  // any temp files
-  fsCollection.files.find(getDoneQuery(fsCollection.options.stores)).observe({
-    added: function(fsFile) {
-      FS.debug && console.log("FileWorker ADDED - calling deleteChunks for", fsFile._id);
-      FS.TempStore.removeFile(fsFile);
-    }
-  });
+
 
   // // Initiate observe for catching files that have been removed and
   // // removing the data from all stores as well
@@ -144,7 +148,7 @@ function getDoneQuery(stores) {
   };
 
   // Add conditions for all defined stores
-  FS.Utility.each(stores, function(store) {
+  FS.Utility.each(stores, function (store) {
     var storeName = store.name;
     var copyCond = {
       $or: [{
